move token helper functions to user_access_token.go
